Migrate snapshotting
